readable guids, flags, ace header flags

+++ pkg/acebuilder.go
@@ -44,12 +44,12 @@ func NewAdvancedAce(buf *bytes.Buffer, totalSize uint16) AdvancedAce {
 	oa := AdvancedAce{}
 	binary.Read(buf, binary.LittleEndian, &oa.Flags)
 	offset := 12
-	if (oa.Flags & uint32(ACEInheritanceFlagsObjectTypePresent)) != 0 {
+	if (oa.Flags & (ACEInheritanceFlagsObjectTypePresent)) != 0 {
 		oa.ObjectType = NewGUID(buf)
 		offset += 16
 	}
 
-	if (oa.Flags & uint32(ACEInheritanceFlagsInheritedObjectTypePresent)) != 0 {
+	if (oa.Flags & (ACEInheritanceFlagsInheritedObjectTypePresent)) != 0 {
 		oa.InheritedObjectType = NewGUID(buf)
 		offset += 16
 	}
